move children components of TopNav & TopNav into TopNav directory
