+++ packages/react-hig/src/elements/components/GlobalNav/TopNav/Profile.test.js
@@ -18,10 +18,10 @@ import { mount } from 'enzyme';
 import * as HIG from 'hig.web';
 import React from 'react';
 
-import GlobalNav from './GlobalNav';
+import GlobalNav from './../GlobalNav';
 import TopNav from './TopNav';
 import Profile from './Profile';
-import SharedExamples from './SharedExamples';
+import SharedExamples from './../SharedExamples';
 
 const onImageClick = function() {
   return 'onImageClick';
