+++ packages/react-hig/src/elements/components/GlobalNav/TopNav/ProjectAccountSwitcher.test.js
@@ -18,10 +18,10 @@ import { mount } from 'enzyme';
 import * as HIG from 'hig.web';
 import React from 'react';
 
-import GlobalNav from './GlobalNav';
+import GlobalNav from './../GlobalNav';
 import TopNav from './TopNav';
 import ProjectAccountSwitcher from './ProjectAccountSwitcher';
-import SharedExamples from './SharedExamples';
+import SharedExamples from './../SharedExamples';
 
 const onAddAccount = function() {
   return 'onAddAcount';
