+++ packages/react-hig/src/elements/components/GlobalNav/TopNav/Shortcut.js
@@ -18,9 +18,9 @@ limitations under the License.
 
  */
 
-import HIGElement from '../../HIGElement';
+import HIGElement from '../../../HIGElement';
 import * as PropTypes from 'prop-types';
-import createComponent from '../../../adapters/createComponent';
+import createComponent from '../../../../adapters/createComponent';
 
 export class Shortcut extends HIGElement {
   commitUpdate(updatePayload, oldProps, newProps) {
