+++ packages/react-hig/src/elements/components/GlobalNav/TopNav/TopNav.test.js
@@ -19,7 +19,7 @@ import { mount } from 'enzyme';
 import * as HIG from 'hig.web';
 import React from 'react';
 
-import GlobalNav from './GlobalNav';
+import GlobalNav from './../GlobalNav';
 import TopNav from './TopNav';
 import Profile from './Profile';
